requirements.txt
transformers==4.36.2
datasets==2.14.5
torch
tqdm
accelerate
fastapi
uvicorn
python-dotenv
openai
numpy